% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{anova_var}
\alias{anova_var}
\title{Analyse de variance (ANOVA) à un facteur}
\usage{
anova_var(data, variable, group)
}
\arguments{
\item{data}{Un data.frame contenant les données.}

\item{variable}{Le nom de la variable quantitative.}

\item{group}{Le nom de la variable catégorielle (facteur).}
}
\value{
Le tableau de résultats de l'ANOVA.
}
\description{
Réalise une ANOVA pour tester l'effet d'un facteur sur une variable quantitative.
}
\examples{
anova_var(iris, "Sepal.Length", "Species")
}
