% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{corr_vars}
\alias{corr_vars}
\title{Matrices de corrélation (Pearson et Spearman)}
\usage{
corr_vars(data, variables)
}
\arguments{
\item{data}{Un data.frame contenant les données.}

\item{variables}{Un vecteur de noms de colonnes à inclure dans les corrélations.}
}
\value{
Une liste contenant deux matrices : pearson et spearman.
}
\description{
Calcule les matrices de corrélation de Pearson et Spearman pour un ensemble de variables quantitatives.
}
\examples{
corr_vars(iris, c("Sepal.Length", "Sepal.Width", "Petal.Length"))
corr_vars(iris, c(1, 2, 3))
}
