% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(mytext, b = 1)
}
\arguments{
\item{mytext}{text to print}
}
\description{
This is the new function. The other function is \code{\link[=littleforecast]{littleforecast()}}.
\code{\link{littleforecast}()}.

Another paragraph.
}
