% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littlefunction.R
\name{littleforecast}
\alias{littleforecast}
\title{Make a ARIMA forecast}
\usage{
littleforecast(data, nyears = 10)
}
\arguments{
\item{data}{This is a time-series of data. Must be a vector
or a ts object.}

\item{nyears}{Number of years in the future to forecast.
Must > 0.}
}
\description{
This function takes a vector or time series of data,
fits a ARIMA function (selected with AIC), and
plots the data. This is different than \link{hello}.
}
\examples{
littleforecast(WWW2)
}
