% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{plot_quanti_quanti}
\alias{plot_quanti_quanti}
\title{Nuage de points entre deux variables quantitatives}
\usage{
plot_quanti_quanti(data, x, y)
}
\arguments{
\item{data}{Un data.frame contenant les données.}

\item{x}{Le nom de la variable en abscisse.}

\item{y}{Le nom de la variable en ordonnée.}
}
\value{
Un objet ggplot2 représentant le graphique.
}
\description{
Produit un graphique représentant la relation entre deux variables quantitatives, accompagné d'une droite de régression linéaire.
}
\examples{
plot_quanti_quanti(iris, "Sepal.Length", "Petal.Length")
}
