% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_foo.R
\name{print.foo}
\alias{print.foo}
\title{Prints foo object}
\usage{
\method{print}{foo}(x, ...)
}
\arguments{
\item{x}{foo object}

\item{...}{Not used}
}
\description{
prints a foo object returned from littlefunction().
}
