% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{stats_var}
\alias{stats_var}
\title{Statistiques descriptives d'une variable quantitative}
\usage{
stats_var(data, variable)
}
\arguments{
\item{data}{Un data.frame contenant les données.}

\item{variable}{Le nom ou numero(position de la colonne) de la variable à analyser}
}
\value{
Un data.frame contenant les statistiques descriptives.
}
\description{
Calcule les principales statistiques descriptives (moyenne, médiane,écart-type, minimum, maximum) pour une variable numérique d'un data.frame.
}
\examples{
stats_var(iris, "Sepal.Length")
stats_var(iris, 1)
}
