% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WWW2}
\alias{WWW2}
\title{Test data: WWW2}
\format{
A data frame with xyz variables
}
\usage{
data(WWW2)
}
\description{
Description
}
\details{
\itemize{
  \item describe the columns
  \item in the dataset
}
}
\keyword{datasets}
