% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littefunction.R
\name{littleforecast}
\alias{littleforecast}
\title{Make a ARIMA forecast}
\usage{
littleforecast()
}
\arguments{
\item{data}{This is a time-series data. Must be a vector or a ts vector}

\item{nyears}{Number of years in the future to forecast.}
}
\description{
This function takes a vector or time series of data,
fits a ARIMA function (selected with AIC), and plots the data
Nerea
}
\examples{
littleforecast (WW2)
}
\references{
REFERENCES.bib
}
\insertRef{Waltonetal1998}{MyNewPackage}
