% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littleforecast.R
\name{littleforecast}
\alias{littleforecast}
\title{Little Forecast}
\usage{
littleforecast(data, nyears = 10, type = c("red", "blue", "green"))
}
\description{
This does an arima forecast
}
\examples{
library(stats)
lm(a ~ b, data=df)
}
