% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-WWW2.R
\docType{data}
\name{WWW2}
\alias{WWW2}
\title{WWW2}
\format{
A ts object.
}
\usage{
data(WWW2)
}
\description{
WWW data squared.

\itemize{
\item price. price in US dollars
\item carat. weight of the diamond
}
}
